WIP move pipe functions to common file
